[add] CartController logic; [add] Unauthenticated exception handling (JSON-404); [add] OrderController logic; [add] Postman testing collection & environment;

+++ project/justbuy-api/app/Http/Controllers/ProductController.php
@@ -58,12 +58,14 @@ class ProductController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Product $product)
+    public function destroy(int $id)
     {
-        $product->delete();
-
-        return response()->json([
-            'message' => 'Product deleted'
-        ], 204);
+        $product = Product::find($id);
+        if ($product) {
+            $product->delete();
+            return response()->json(['message' => 'Product deleted']);
+        } else {
+            return response()->json(['message' => 'Product not found'], 404);
+        }
     }
 }
